added a tool to measure the performance of a specific version of hibridon
This benchmarking tool includes `starbench` a benchmarking method (inspired by `hpl`'s `stardgemm`) which runs the same command in parallel, so that performance is measured in conditions similar to what a user would experience in a compute cluster environment.

This is functional but still work in progress (the code dates from the 03/05/2022).